@@ -79,10 +79,13 @@ const (
 	organizationFlag       = "org"
 	organizationalUnitFlag = "ou"
 	outFlag                = "out"
+	separateOutFlag        = "separate"
 	passwordFlag           = "pass"
 	postalCodeFlag         = "postalcode"
 	provinceFlag           = "province"
 	rootOutFlag            = "rootout"
+	rootsOnlyFlag          = "roots"
+	intermediatesOnlyFlag  = "intermediates"
 	separatorFlag          = "separator"
 	serialNumberFlag       = "sn"
 	serverFlag             = "server"
@@ -204,6 +207,10 @@ var optDefs = map[string]option{
 		desc:         "output file",
 		defaultValue: "",
 	},
+	separateOutFlag: {
+		desc:         fmt.Sprintf("write every CA certificate to a separate file with optional prefix specified by -%s", outFlag),
+		defaultValue: false,
+	},
 	passwordFlag: {
 		argFmt:       stringFmt,
 		defaultLabel: "none",
@@ -211,8 +218,15 @@ var optDefs = map[string]option{
 		defaultValue: "",
 	},
 	rootOutFlag: {
-		argFmt:       pathFmt,
-		desc:         "output root CA certificate only",
+		desc:         fmt.Sprintf("output the first root CA certificate only; this flag will be deprecated in favor of -%s, which is preferred", rootsOnlyFlag),
+		defaultValue: false,
+	},
+	rootsOnlyFlag: {
+		desc:         "only output root (self-signed) certificates",
+		defaultValue: false,
+	},
+	intermediatesOnlyFlag: {
+		desc:         "only output intermediate certificates",
 		defaultValue: false,
 	},
 	separatorFlag: {
